requirements.txt
Django==3.0
pytz==2018.5
djangorestframework==3.14.0
psycopg2==2.9.5
